FahrzeugPanel-Content aufgeteilt
