+++ frontend/components.d.ts
@@ -31,6 +31,8 @@ declare module '@vue/runtime-core' {
     DetektierteFahrzeugartIcon: typeof import('./src/components/messstelle/DetektierteFahrzeugartIcon.vue')['default']
     DisplayImageAsset: typeof import('./src/components/pdfreport/assets/DisplayImageAsset.vue')['default']
     FahrzeugPanel: typeof import('./src/components/messstelle/optionsmenue/panels/FahrzeugPanel.vue')['default']
+    FahrzeugPanelFahrzeugkategorienContent: typeof import('./src/components/messstelle/optionsmenue/panels/content/FahrzeugPanelFahrzeugkategorienContent.vue')['default']
+    FahrzeugPanelVerkehrsartenContent: typeof import('./src/components/messstelle/optionsmenue/panels/content/FahrzeugPanelVerkehrsartenContent.vue')['default']
     FilterOptionen: typeof import('./src/components/messstelle/optionsmenue/FilterOptionen.vue')['default']
     GeometriePanel: typeof import('./src/components/zaehlstelle/optionsmenue/panels/GeometriePanel.vue')['default']
     HeadingAssetForm: typeof import('./src/components/pdfreport/assetforms/HeadingAssetForm.vue')['default']
